requirements.txt
requests~=2.32.3
python-dotenv~=1.0.1
woocommerce~=3.0.0
